Add src file to organize routes and tests
